Adjustments for `bot` structure refactor
